Replace using arbitrary proxy with no address proxy with no address
