+++ src/relay/relay.py
@@ -45,7 +45,7 @@ from .blockchain.events import BlockchainEvent
 from .blockchain.events_informations import EventsInformationFetcher
 from .blockchain.exchange_proxy import ExchangeProxy
 from .blockchain.node import Node
-from .blockchain.proxy import sorted_events
+from .blockchain.proxy import Proxy, sorted_events
 from .blockchain.token_proxy import TokenProxy
 from .blockchain.unw_eth_proxy import UnwEthProxy
 from .events import BalanceEvent, NetworkBalanceEvent
@@ -162,7 +162,7 @@ class TrustlinesRelay:
     def use_eth_index(self) -> bool:
         return os.environ.get("ETHINDEX", "1") == "1"
 
-    def get_event_selector_for_currency_network(self, network_address):
+    def get_event_selector_for_currency_network(self, network_address=None):
         """return either a CurrencyNetworkProxy or a EthindexDB instance
         This is being used from relay.api to query for events.
         """
@@ -175,22 +175,9 @@ class TrustlinesRelay:
                 from_to_types=currency_network_events.from_to_types,
             )
         else:
+            if network_address is None:
+                return Proxy(self._web3, abi=self.contracts["CurrencyNetwork"]["abi"])
             return self.currency_network_proxies[network_address]
-
-    def get_event_selector_without_currency_network(self):
-        """Returns either a CurrencyNetworkProxy or a EthindexDB instance
-        This can be used to query events where address is unknown
-        or via providing the missing address"""
-        if not self.use_eth_index:
-            # pick arbitrarily a proxy in the dict
-            return list(self.currency_network_proxies.values())[0]
-        return ethindex_db.EthindexDB(
-            ethindex_db.connect(""),
-            address=None,
-            standard_event_types=currency_network_events.standard_event_types,
-            event_builders=currency_network_events.event_builders,
-            from_to_types=currency_network_events.from_to_types,
-        )
 
     def get_event_selector_for_token(self, address):
         """return either a proxy or a EthindexDB instance
@@ -249,6 +236,9 @@ class TrustlinesRelay:
     def get_network_info(self, network_address: str) -> NetworkInfo:
         proxy = self.currency_network_proxies[network_address]
         graph = self.currency_network_graphs[network_address]
+        assert (
+            proxy.address is not None
+        ), "Invalid currency network proxy with no address."
         return NetworkInfo(
             address=proxy.address,
             name=proxy.name,
@@ -294,7 +284,7 @@ class TrustlinesRelay:
 
     def get_transfer_information(self, tx_hash):
         fetcher = EventsInformationFetcher(
-            self.get_event_selector_without_currency_network()
+            self.get_event_selector_for_currency_network()
         )
         return fetcher.get_transfer_details(tx_hash)
